editing delete and app.blede d

+++ app/Modules/Admin/Http/Controllers/RestaurantController.php
@@ -39,28 +39,35 @@ class RestaurantController extends Controller
     {
         if ($request->type == 'image') {
             $image = Image::find($id);
-            if (file_exists(substr($image->path, 1))) {
-                unlink(substr($image->path, 1));
-            }
+            $path = $image->path;
             Image::destroy($id);
+            if (file_exists(substr($path, 1))&& $path!='/img/default/rest.jpg') {
+                unlink(substr($path, 1));
+            }
         } elseif ($request->type == 'menu') {
             $doc = Document::find($id);
-            if (file_exists(substr($doc->path, 1))) {
+            $path = $doc->path;
+            Document::destroy($id);
+
+            if (file_exists(substr($doc->path, 1)) && $path!='/img/default/rest.jpg') {
                 unlink(substr($doc->path, 1));
             }
-            Document::destroy($id);
         }
         if ($request->type == 'address') {
             Address::destroy(substr($image->path, 1));
         }
     }
-    public function hiddeRest(Request $request, $id){
-        $restaurant = Restaurant::find($id);
-        if($request->visible == true){
-            $restaurant->visible=1;
+
+    public function hiddeRest(Request $request, $id)
+    {
+        $restaurant = Restaurant::find($id)->first();
+
+        if ($request->visible == true) {
+            $restaurant->visible = 1;
             $restaurant->save();
-        }else{
-            $restaurant->visible=0;
+        } else {
+
+            $restaurant->visible = 0;
             $restaurant->save();
         }
 
@@ -99,7 +106,28 @@ class RestaurantController extends Controller
                     'lng' => $address['lng'],
                     'restaurant_id' => $restaurant->id]);
             }
+        $this->validateUpdate($id);
         return redirect('/');
+
+    }
+
+    public function validateUpdate($id)
+    {
+        $image = Image::where('restaurant_id',$id)->first();
+        if (!count($image)) {
+           Image::create([
+               'path'=>'/img/default/rest.jpg',
+               'restaurant_id'=>$id,
+           ]);
+        }
+        $doc= Document::where('restaurant_id',$id)->first();
+        if (!count($doc)) {
+           Document::create([
+               'type'=>'image',
+               'path'=>'/img/default/rest.jpg',
+               'restaurant_id'=>$id,
+           ]);
+        }
 
     }
